Major Bug Fixes
cleanup active replicators on database close
changed implementation of transactions to work on more recent android sdks
properly catch sqlexception and return error code in forceInsert
sort list of databases returned for consistent behavior
recognize -1 as TDStatus in error condition
avoid NPE in TDView
improved shutdown behavior of TDPuller
fixed bug in TDPusher extracting source URL from changes
added handler thread to TDReplicator
synchronized methods with concurrency issue
added cleanup call to batcher
fixed error handling when remote checkpoint returns 404
added handler thread to change tracker
added code to abort request when shutting down
fixed npe in router cleanup
fixed bug in chunked response for changes feed
added flag for chunked response in TDURLConnection
renamed TDURLStreamHandlerFactor to TDURLStreamHandlerFactory
added handler thread to batcher
added cleanup method to batcher
added error logging to TDRemoteRequest
properly cleanup handler thread in TDRemoteRequest

+++ TouchDB-Android/src/com/couchbase/touchdb/router/TDRouter.java
@@ -359,6 +359,7 @@ public class TDRouter implements Observer {
             }
         } catch (Exception e) {
             //default status is internal server error
+            Log.e(TDDatabase.TAG, "Exception in TDRouter", e);
         }
 
         // Configure response headers:
@@ -397,7 +398,9 @@ public class TDRouter implements Observer {
 
     public void stop() {
         callbackBlock = null;
-        db.deleteObserver(this);
+        if(db != null) {
+            db.deleteObserver(this);
+        }
     }
 
     public TDStatus do_UNKNOWN(TDDatabase db, String docID, String attachmentName) {
@@ -806,9 +809,10 @@ public class TDRouter implements Observer {
         boolean continuous = !longpoll && "continuous".equals(feed);
 
         if(continuous || (longpoll && changes.size() == 0)) {
+            connection.setChunked(true);
+            connection.setResponseCode(TDStatus.OK);
+            sendResponse();
             if(continuous) {
-                connection.setResponseCode(TDStatus.OK);
-                sendResponse();
                 for (TDRevision rev : changes) {
                     sendContinuousChange(rev);
                 }
